# Conflicts:
#	core/src/main/java/io/questdb/PropServerConfiguration.java
#	core/src/main/java/io/questdb/ServerMain.java
#	core/src/main/java/io/questdb/cutlass/http/processors/ExportQueryProcessor.java
#	core/src/main/java/io/questdb/cutlass/text/CopyImportJob.java
#	core/src/test/java/io/questdb/test/ServerMainTest.java
#	core/src/test/java/io/questdb/test/cutlass/http/HttpErrorHandlingTest.java
#	core/src/test/java/io/questdb/test/cutlass/pgwire/BasePGTest.java
#	core/src/test/java/io/questdb/test/cutlass/pgwire/PGErrorHandlingTest.java
#	core/src/test/java/io/questdb/test/griffin/AlterTableConvertPartitionTest.java
#	core/src/test/java/io/questdb/test/griffin/CopyImportTest.java

+++ compat/src/test/java/io/questdb/compat/InfluxDBClientFailureTest.java
@@ -28,7 +28,6 @@ import io.questdb.Bootstrap;
 import io.questdb.DefaultBootstrapConfiguration;
 import io.questdb.DefaultHttpClientConfiguration;
 import io.questdb.ServerMain;
-import io.questdb.cairo.CairoException;
 import io.questdb.cutlass.http.client.HttpClient;
 import io.questdb.cutlass.http.client.HttpClientFactory;
 import io.questdb.griffin.SqlException;
@@ -55,7 +54,7 @@ public class InfluxDBClientFailureTest extends AbstractTest {
             private final AtomicInteger attempt = new AtomicInteger();
 
             @Override
-            public long openRW(LPSZ name, long opts) {
+            public long openRW(LPSZ name, int opts) {
                 if (Utf8s.endsWithAscii(name, Files.SEPARATOR + "x.d") && attempt.getAndIncrement() == 0) {
                     return -1;
                 }
@@ -95,7 +94,7 @@ public class InfluxDBClientFailureTest extends AbstractTest {
             private final AtomicInteger attempt = new AtomicInteger();
 
             @Override
-            public long openRW(LPSZ name, long opts) {
+            public long openRW(LPSZ name, int opts) {
                 if (Utf8s.endsWithAscii(name, Files.SEPARATOR + "x.d") && attempt.getAndIncrement() == 0) {
                     throw new OutOfMemoryError();
                 }
@@ -135,7 +134,7 @@ public class InfluxDBClientFailureTest extends AbstractTest {
             private final AtomicInteger counter = new AtomicInteger(1);
 
             @Override
-            public long openRW(LPSZ name, long opts) {
+            public long openRW(LPSZ name, int opts) {
                 if (Utf8s.endsWithAscii(name, Files.SEPARATOR + EVENT_INDEX_FILE_NAME)
                         && Utf8s.containsAscii(name, "failed_table")
                         && (counter.getAndDecrement() > 0)) {
@@ -180,7 +179,7 @@ public class InfluxDBClientFailureTest extends AbstractTest {
             private final AtomicInteger attempt = new AtomicInteger();
 
             @Override
-            public long openRW(LPSZ name, long opts) {
+            public long openRW(LPSZ name, int opts) {
                 if (Utf8s.endsWithAscii(name, Files.SEPARATOR + "x.d") && attempt.getAndIncrement() == 0) {
                     try {
                         server.get().getEngine().execute("drop table m1");
@@ -220,7 +219,7 @@ public class InfluxDBClientFailureTest extends AbstractTest {
             private final AtomicInteger attempt = new AtomicInteger();
 
             @Override
-            public long openRW(LPSZ name, long opts) {
+            public long openRW(LPSZ name, int opts) {
                 if (Utf8s.endsWithAscii(name, Files.SEPARATOR + "x.d") && attempt.getAndIncrement() == 0) {
                     try {
                         server.get().getEngine().execute("drop table m1");
@@ -288,7 +287,7 @@ public class InfluxDBClientFailureTest extends AbstractTest {
             private final AtomicInteger counter = new AtomicInteger(0);
 
             @Override
-            public long openRW(LPSZ name, long opts) {
+            public long openRW(LPSZ name, int opts) {
                 if (Utf8s.endsWithAscii(name, Files.SEPARATOR + "z.d") && counter.getAndIncrement() == 0) {
                     throw new UnsupportedOperationException();
                 }
@@ -337,7 +336,7 @@ public class InfluxDBClientFailureTest extends AbstractTest {
             private final AtomicInteger attempt = new AtomicInteger();
 
             @Override
-            public long openRW(LPSZ name, long opts) {
+            public long openRW(LPSZ name, int opts) {
                 if (Utf8s.endsWithAscii(name, Files.SEPARATOR + "good_y.d") && attempt.getAndIncrement() == 0) {
                     try {
                         server.get().getEngine().execute("drop table \"drop\"");
